+++ src/main/java/org/kdb/inside/brains/psi/QPsiUtil.java
@@ -5,7 +5,6 @@ import com.intellij.openapi.editor.ScrollType;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.TextRange;
 import com.intellij.openapi.util.text.StringUtil;
-import com.intellij.psi.*;
 import com.intellij.psi.impl.source.tree.LeafPsiElement;
 import com.intellij.psi.util.PsiEditorUtil;
 import com.intellij.psi.util.PsiTreeUtil;
@@ -14,8 +13,9 @@ import org.jetbrains.annotations.Nullable;
 import org.kdb.inside.brains.QFileType;
 
 import java.util.ArrayList;
-import java.util.Collection;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.function.Predicate;
 
 public final class QPsiUtil {
@@ -101,20 +101,59 @@ public final class QPsiUtil {
         };
     }
 
-    public static boolean isInnerDeclaration(QLambda lambda, String name) {
+
+    public static Map<String, QVarDeclaration> getLambdaDeclarations(@NotNull QLambda lambda) {
+        final Map<String, QVarDeclaration> res = new HashMap<>();
+        lambda.acceptChildren(new PsiRecursiveElementWalkingVisitor() {
+            @Override
+            public void visitElement(@NotNull PsiElement element) {
+                // we don't search inside others lambda, query, table or dict
+                if (isPrivateContextHolder(element)) {
+                    return;
+                }
+
+                if (element instanceof QVarDeclaration d) {
+                    if (!QPsiUtil.isGlobalDeclaration(d)) {
+                        res.put(d.getName(), d);
+                    }
+                } else {
+                    super.visitElement(element);
+                }
+            }
+        });
+        return res;
+    }
+
+    public static @Nullable QVarDeclaration getLocalDefinition(@NotNull QLambda lambda, @NotNull QVariable variable) {
+        final String name = variable.getName();
         // implicit variable or in parameters list - ignore namespace
         if (lambda.getParameters() == null && QPsiUtil.isImplicitName(name)) {
-            return true;
+            return null;
         }
 
-        final Collection<QVarDeclaration> declarations = PsiTreeUtil.findChildrenOfType(lambda, QVarDeclaration.class);
-        for (QVarDeclaration declaration : declarations) {
-            // Same name and same lambda - internal variable
-            if (name.equals(declaration.getName()) && !QPsiUtil.isGlobalDeclaration(declaration)) {
-                return true;
+        final QVarDeclaration[] res = new QVarDeclaration[1];
+        lambda.acceptChildren(new PsiRecursiveElementWalkingVisitor() {
+            @Override
+            public void visitElement(@NotNull PsiElement element) {
+                // we don't search inside others lambda, query, table or dict
+                if (isPrivateContextHolder(element)) {
+                    return;
+                }
+
+                // no reason to search after the definition.// In a lambda it can't be below.
+                if (element.equals(variable)) {
+                    stopWalking();
+                } else if (element instanceof QVarDeclaration d) {
+                    if (name.equals(d.getName()) && !QPsiUtil.isGlobalDeclaration(d)) {
+                        res[0] = d;
+                        stopWalking();
+                    }
+                } else {
+                    super.visitElement(element);
+                }
             }
-        }
-        return false;
+        });
+        return res[0];
     }
 
     /**
@@ -342,5 +381,9 @@ public final class QPsiUtil {
 
     public static QVarDeclaration createVarDeclaration(Project project, String name) {
         return PsiTreeUtil.findChildOfType(QFileType.createFactoryFile(project, name + ":`"), QVarDeclaration.class);
+    }
+
+    public static boolean isPrivateContextHolder(@NotNull PsiElement element) {
+        return element instanceof QLambdaExpr || element instanceof QQueryExpr || element instanceof QTableExpr || element instanceof QDictExpr;
     }
 }