+++ src/main/java/org/kdb/inside/brains/psi/refs/QSpecReferenceProvider.java
@@ -92,7 +92,7 @@ public class QSpecReferenceProvider extends QBaseReferenceProvider<QVariable> {
         }
 
         @Override
-        protected ResolveResult[] resolveElement(QVariable element) {
+        protected ResolveResult[] resolveGlobalDefinition(QVariable element) {
             final TestDescriptor desc = TestDescriptor.of(element);
             if (desc == null) {
                 return ResolveResult.EMPTY_ARRAY;